src/requirements.txt
Flask
Flask-RESTful
Flask-CORS
Flask-SQLAlchemy
Flask-Session
Flask-Dance
oauthlib
requests
mysql-connector-python
